back/requirements.txt
flask==2.2.2
flask-cors==3.0.10
gunicorn==20.1.0
google-api-core
google-cloud-compute